Have Legacy Mode use a Popup Window
Starting with Chrome 94, we can no longer use Ajax requests to
communicate via http with the Companion. We also cannot communicate
with http if we are in a secure context.

This change introduces a Popup window which is loaded from the
Companion itself and can therefore communicate with the Companion. Our
main window (from replmgr.js) will now communicate with this Popup
which will act as a go between. This not only permits Chrome 94+ to
work but also permits us to load App Inventor in a secure context.

This code is both forward and backward compatible. An older Companion
will use Legacy Mode as it always has (and be subject to the
restrictions involved). Similarly the new Companion will work with an
older version of App Inventor.

The new Companion indicates that it can use a popup by sending the
"useproxy" flag to the Rendezvous server which is then queried for in
replmgr.js

Change-Id: I43005b78c190dfb8446308f692dc39dec51befb1